+++ src/features/discussions/ui/vote/Vote.tsx
